lora-php-ephect/requirements.txt
torch
torchvision
torchaudio
transformers
datasets
accelerate
peft
bitsandbytes
trl
sentencepiece
pyyaml
tqdm
